chore(refactor): Refactoring the wait logic out of blockingproxy.ts.
